requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
email-validator
python-dotenv==1.0.0
WTForms==3.0.1
Werkzeug==2.3.7
itsdangerous==2.1.2
pymysql==1.1.0
cryptography
requests==2.31.0
Flask-Limiter==3.5.0 